pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-qa-tester"
version = "0.1.0"
description = "AI-powered QA code tester scaffold"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "httpx>=0.27,<1.0",
  "pillow>=10.4,<12.0",
  "opencv-python-headless>=4.10,<5.0",
  "pytesseract>=0.3.13,<1.0"
]

[project.optional-dependencies]
runner = [
  "playwright>=1.49,<2.0"
]
cloud = [
  "azure-servicebus>=7.13,<8.0",
  "azure-storage-blob>=12.23,<13.0",
  "qdrant-client>=1.11,<2.0",
  "psycopg[binary]>=3.2,<4.0"
]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.24,<1.0",
  "ruff>=0.6,<1.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
